Built and trained model m5
Model m5 is a U-net architecture with a dense middle layer. Improvements to this model include adding a better activation function for the outputs, and possibly an attention layer. For further training, it will be possible to include the predictions of the model in the dataset. This will allow the model to learn how to optimize its own data.